refactor(ui): extract workspace close context helper
